LangGraph/acf-integration/requirements.txt
fastapi
pydantic
uvicorn
python-dotenv
ibm-watsonx-ai 
langchain-core
langchain-ibm
langgraph
pymilvus
tavily